% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art02_and_26_assistance_provided}
\alias{art02_and_26_assistance_provided}
\title{Article 2_and_26_assistance_provided}
\format{A dataset with 3042 rows and 4 columns.}
\usage{
art02_and_26_assistance_provided
}
\description{
Article 2_and_26_assistance_provided
}
\details{
This article includes the following sections:
\itemize{
\item 4.1 - On transfer of skills and technology (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4710}{website})
\item 4.2 - Expertise for tobacco control programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4713}{website})
\item 4.3 - Training and sensitisation of personnel (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4715}{website})
\item 4.4 - Equipment, supplies, logistics (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4717}{website})
\item 4.5 - Methods for tobacco control, e.g. treatment of nicotine addiction (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4719}{website})
\item 4.6 - Research on affordability (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5744}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art16}}, \code{\link{art17}},
  \code{\link{art18}}, \code{\link{art19}},
  \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
