% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art02_and_26_other_questions_related_to_assistance}
\alias{art02_and_26_other_questions_related_to_assistance}
\title{Article 2_and_26_other_questions_related_to_assistance}
\format{A dataset with 2535 rows and 4 columns.}
\usage{
art02_and_26_other_questions_related_to_assistance
}
\description{
Article 2_and_26_other_questions_related_to_assistance
}
\details{
This article includes the following sections:
\itemize{
\item 4.7 - Identify the Party or Parties (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4729}{website})
\item 4.8 - Information about assistance (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5847}{website})
\item 4.9 - Assistance under consideration (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4727}{website})
\item 4.10 - Encouraging development institutions (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5844}{website})
\item 4.11 - Additional details on encouraging development institutions (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5845}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art16}}, \code{\link{art17}},
  \code{\link{art18}}, \code{\link{art19}},
  \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
