% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art02_and_26_priorities_and_comments}
\alias{art02_and_26_priorities_and_comments}
\title{Article 2_and_26_priorities_and_comments}
\format{A dataset with 3042 rows and 4 columns.}
\usage{
art02_and_26_priorities_and_comments
}
\description{
Article 2_and_26_priorities_and_comments
}
\details{
This article includes the following sections:
\itemize{
\item 5.1 - Priorities for implementation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4487}{website})
\item 5.2 - Specific gaps (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5745}{website})
\item 5.3 - Details on specific gaps (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5747}{website})
\item 5.4 - Constraints or barriers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4488}{website})
\item 5.5 - Any other relevant information (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4489}{website})
\item 5.6 - Suggestions on the reporting instrument (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5746}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art16}}, \code{\link{art17}},
  \code{\link{art18}}, \code{\link{art19}},
  \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
