% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art05}
\alias{art05}
\title{Article 5}
\format{A dataset with 8619 rows and 4 columns.}
\usage{
art05
}
\description{
Article 5
}
\details{
This article includes the following sections:
\itemize{
\item 3.1.1.1 - Comprehensive multisectoral national tobacco control strategy (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5048}{website})
\item 3.1.1.2 - Tobacco control included in national health strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5049}{website})
\item 3.1.1.3 - Any aspect of tobacco referred in any national strategy (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5050}{website})
\item 3.1.1.4 - Focal point for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5784}{website})
\item 3.1.1.4 - Tobacco control unit (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5212}{website})
\item 3.1.1.4 - Tobacco control unit (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5212}{website})
\item 3.1.1.4 - National coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5785}{website})
\item 3.1.1.4 - National coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5785}{website})
\item 3.1.1.5 - Details on focal point for tobacco control, tobacco control unit and national coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5214}{website})
\item 3.1.1.6 - Progress made in implementing Article 5.1 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5215}{website})
\item 3.1.1.7 - Additional information concerning general obligations (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5216}{website})
\item 3.1.2.1 - Interference by the tobacco industry (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5067}{website})
\item 3.1.2.2 - Public access to a wide range of information on the tobacco industry (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5790}{website})
\item 3.1.2.3 - Details on the protection of public health policies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5787}{website})
\item 3.1.2.4 - Progress made in protection of public health policies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5788}{website})
\item 3.1.2.5 - Use of guidelines on Article 5.3 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7004}{website})
\item 3.1.2.6 - Additional information concerning protection of public health policies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5789}{website})
\item 3.1.1.4 - Tobacco control unit (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5212}{website})
\item 3.1.1.4 - Tobacco control unit (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5212}{website})
\item 3.1.1.4 - National coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5785}{website})
\item 3.1.1.4 - National coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5785}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art06}}, \code{\link{art08}},
  \code{\link{art09}}, \code{\link{art10}},
  \code{\link{art11}}, \code{\link{art12}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
