% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art06}
\alias{art06}
\title{Article 6}
\format{A dataset with 8619 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art06
}
\description{
Article 6
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 2.3.1 - Existence of information on tobacco-related mortality (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5528}{website})
\item 2.3.3 - Additional information on tobacco-related mortality (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5530}{website})
\item 2.4.1 - Existence of information on the economic burden of tobacco use (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5534}{website})
\item 2.8.1 - Proportion of the retail price consisting of taxes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5484}{website})
\item 2.8.1 - Proportion of the retail price consisting of taxes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5484}{website})
\item 2.8.2 - Only specific tax levied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5485}{website})
\item 2.8.2 - Only ad valorem tax levied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5486}{website})
\item 2.8.2 - Combination of specific and ad valorem taxes levied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5487}{website})
\item 2.8.2 - Complex tax structure applied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5488}{website})
\item 2.8.4 - Trends in taxation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5502}{website})
\item 2.8.5 - Tobacco tax earmarking (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5503}{website})
\item 2.9.4 - Trends in prices (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5524}{website})
\item 3.2.1.1 - Tax policies to reduce tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5222}{website})
\item 3.2.1.2 - Tobacco sales to international travelers prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5223}{website})
\item 3.2.1.3 - Tobacco imports by international travelers prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5224}{website})
\item 3.2.1.4 - Progress made in implementing Article 6 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5225}{website})
\item 3.2.1.5 - Additional information concerning price and tax measures (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5226}{website})
\item 2.8.1 - Proportion of the retail price consisting of taxes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5484}{website})
\item 2.8.1 - Proportion of the retail price consisting of taxes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5484}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art08}},
  \code{\link{art09}}, \code{\link{art10}},
  \code{\link{art11}}, \code{\link{art12}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
