% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art08}
\alias{art08}
\title{Article 8}
\format{A dataset with 7098 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art08
}
\description{
Article 8
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 2.2.1 - Availability of data on exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5525}{website})
\item 2.2.2 - Details concerning data on exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5526}{website})
\item 3.2.2.1 - Tobacco smoking banned in public places (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5026}{website})
\item 3.2.2.2 - National law providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5851}{website})
\item 3.2.2.2 - Subnational law(s) providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5852}{website})
\item 3.2.2.2 - Administrative and executive orders providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5853}{website})
\item 3.2.2.2 - Voluntary agreements providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5854}{website})
\item 3.2.2.3 - Explanation of type/nature and content of the measures providing the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5857}{website})
\item 3.2.2.4 - Mechanism/infrastructure for enforcement provided (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5858}{website})
\item 3.2.2.5 - Details of the mechanism/infrastructure for enforcement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5859}{website})
\item 3.2.2.5 - Details of the mechanism/infrastructure for enforcement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5859}{website})
\item 3.2.2.8 - Progress made in implementing Article 8 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5570}{website})
\item 3.2.2.9 - Use of guidelines on Article 8 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7009}{website})
\item 3.2.2.10 - Additional information concerning protection from exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5571}{website})
\item 3.2.2.5 - Details of the mechanism/infrastructure for enforcement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5859}{website})
\item 3.2.2.5 - Details of the mechanism/infrastructure for enforcement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5859}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art09}}, \code{\link{art10}},
  \code{\link{art11}}, \code{\link{art12}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
