% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art09}
\alias{art09}
\title{Article 9}
\format{A dataset with 3549 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art09
}
\description{
Article 9
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 3.2.3.1 - Testing and measuring the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4430}{website})
\item 3.2.3.2 - Testing and measuring the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4431}{website})
\item 3.2.3.3 - Regulating the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4432}{website})
\item 3.2.3.4 - Regulating the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4433}{website})
\item 3.2.3.5 - Progress made in implementing Article 9 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5547}{website})
\item 3.2.3.6 - Use of guidelines on Article 9 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7011}{website})
\item 3.2.3.7 - Additional information concerning regulation of the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5548}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art10}},
  \code{\link{art11}}, \code{\link{art12}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
