% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art10}
\alias{art10}
\title{Article 10}
\format{A dataset with 4056 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art10
}
\description{
Article 10
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 3.2.4.1 - Requiring disclosure of information about the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5043}{website})
\item 3.2.4.1 - Requiring disclosure of information about the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5044}{website})
\item 3.2.4.2 - Requiring public disclosure on the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5791}{website})
\item 3.2.4.2 - Requiring public disclosure on the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5791}{website})
\item 3.2.4.2 - Requiring public disclosure on the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5792}{website})
\item 3.2.4.3 - Progress made in implementing Article 10 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5793}{website})
\item 3.2.4.5 - Additional information concerning regulation of tobacco product disclosures (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5794}{website})
\item 3.2.4.4 - Use of guidelines on Article 10 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7012}{website})
\item 3.2.4.2 - Requiring public disclosure on the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5791}{website})
\item 3.2.4.2 - Requiring public disclosure on the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5791}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art11}}, \code{\link{art12}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
