% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art11}
\alias{art11}
\title{Article 11}
\format{A dataset with 9633 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art11
}
\description{
Article 11
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 3.2.5.1 - Packaging of tobacco products does not carry advertisement or promotion (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5575}{website})
\item 3.2.5.2 - Misleading descriptors (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4434}{website})
\item 3.2.5.3 - Health warnings required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4435}{website})
\item 3.2.5.4 - Health warnings approved by the competent national authority (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4436}{website})
\item 3.2.5.5 - Rotated health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4437}{website})
\item 3.2.5.6 - Large, clear, visible and legible health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4438}{website})
\item 3.2.5.8 - Health warnings occupying no less than 30% (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4439}{website})
\item 3.2.5.8 - Health warnings occupying no less than 30% (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4439}{website})
\item 3.2.5.9 - Health warnings occupying 50% or more (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4440}{website})
\item 3.2.5.9 - Health warnings occupying 50% or more (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4440}{website})
\item 3.2.5.10 - Health warnings in the form of pictures or pictograms (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4441}{website})
\item 3.2.5.11 - Copyright to pictures owned by the Government (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5576}{website})
\item 3.2.5.12 - Granting of license for the use of health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5578}{website})
\item 3.2.5.12 - Granting of license for the use of health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5578}{website})
\item 3.2.5.13 - Constituents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5874}{website})
\item 3.2.5.13 - Emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5875}{website})
\item 3.2.5.15 - Progress made in implementing Article 11 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5579}{website})
\item 3.2.5.16 - Use of guidelines on Article 11 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7018}{website})
\item 3.2.5.17 - Additional information concerning packaging and labelling (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5580}{website})
\item 3.2.5.8 - Health warnings occupying no less than 30% (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4439}{website})
\item 3.2.5.8 - Health warnings occupying no less than 30% (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4439}{website})
\item 3.2.5.9 - Health warnings occupying 50% or more (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4440}{website})
\item 3.2.5.9 - Health warnings occupying 50% or more (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4440}{website})
\item 3.2.5.12 - Granting of license for the use of health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5578}{website})
\item 3.2.5.12 - Granting of license for the use of health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5578}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art12}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
