% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art12}
\alias{art12}
\title{Article 12}
\format{A dataset with 16731 rows and 4 columns.}
\usage{
art12
}
\description{
Article 12
}
\details{
This article includes the following sections:
\itemize{
\item 3.2.6.1 - Implemented educational and public awareness programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5069}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to adults (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5070}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to children and youth (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5071}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to men (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5583}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5585}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to pregnant women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5587}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to ethnic groups (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5588}{website})
\item 3.2.6.3 - Age differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5594}{website})
\item 3.2.6.3 - Gender differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5595}{website})
\item 3.2.6.3 - Educational background differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5596}{website})
\item 3.2.6.3 - Cultural differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5597}{website})
\item 3.2.6.3  - Socioeconomic status reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5599}{website})
\item 3.2.6.4 - Programmes covering the health risks of tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5072}{website})
\item 3.2.6.4 - Programmes covering the risks of exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5073}{website})
\item 3.2.6.4 - Programmes covering the benefits of cessation of tobacco use (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5074}{website})
\item 3.2.6.4 - Programmes covering economic consequences of tobacco production (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5601}{website})
\item 3.2.6.4 - Programmes covering economic consequences of tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5604}{website})
\item 3.2.6.4 - Programmes covering environmental consequences of tobacco production (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5606}{website})
\item 3.2.6.4 - Programmes covering environmental consequences of tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5608}{website})
\item 3.2.6.5 - Public agencies involved in programmes and strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5611}{website})
\item 3.2.6.5 - NGOs involved in programmes and strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5612}{website})
\item 3.2.6.5 - Private organizations involved in programmes and strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5615}{website})
\item 3.2.6.6 - Programmes guided by research (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5618}{website})
\item 3.2.6.7 - Training programmes addressed to health workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5090}{website})
\item 3.2.6.7 - Training programmes addressed to community workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5091}{website})
\item 3.2.6.7 - Training programmes addressed to social workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5092}{website})
\item 3.2.6.7 - Training programmes addressed to media professionals (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5093}{website})
\item 3.2.6.7 - Training programmes addressed to educators (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5094}{website})
\item 3.2.6.7 - Training programmes addressed to decision-makers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5095}{website})
\item 3.2.6.7 - Training programmes addressed to administrators (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5096}{website})
\item 3.2.6.8 - Progress made in implementing Article 12 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5621}{website})
\item 3.2.6.9 - Use of guidelines on Article 12 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7013}{website})
\item 3.2.6.10 - Additional information concerning education and communication (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5622}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art13}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
