% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art13}
\alias{art13}
\title{Article 13}
\format{A dataset with 17745 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art13
}
\description{
Article 13
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 3.2.7.1 - Comprehensive ban on all tobacco advertising, promotion and sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5346}{website})
\item 3.2.7.2 - Ban on display of tobacco products at points of sales (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5349}{website})
\item 3.2.7.2 - Ban on display of tobacco products at points of sales (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5349}{website})
\item 3.2.7.2 - Ban covering the domestic internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5350}{website})
\item 3.2.7.2 - Ban covering the global internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5351}{website})
\item 3.2.7.2 - Ban covering brand stretching and/or sharing (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5352}{website})
\item 3.2.7.2 - Ban covering product placement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5353}{website})
\item 3.2.7.2 - Ban covering product placement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5353}{website})
\item 3.2.7.2 - Ban covering the depiction/use of tobacco in entertainment media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5354}{website})
\item 3.2.7.2 - Ban covering the depiction/use of tobacco in entertainment media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5354}{website})
\item 3.2.7.2 - Ban covering tobacco sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5356}{website})
\item 3.2.7.2 - Ban covering tobacco sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5356}{website})
\item 3.2.7.2 - Ban covering corporate social responsibility (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5357}{website})
\item 3.2.7.2 - Ban covering cross-border advertising originating from the country (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5358}{website})
\item 3.2.7.2 - Ban covering cross-border advertising entering the country (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5359}{website})
\item 3.2.7.3 - Precluded by constitution from undertaking a comprehensive ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5360}{website})
\item 3.2.7.4 - All tobacco advertising, promotion and sponsorship restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5361}{website})
\item 3.2.7.5 - Cross-border advertising originating from the country restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5362}{website})
\item 3.2.7.6 - Advertising by false and misleading means prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5363}{website})
\item 3.2.7.7 - Use of warnings to accompany all advertising required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5364}{website})
\item 3.2.7.8 - Use of direct or indirect incentives restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5365}{website})
\item 3.2.7.9 - Disclosure of advertising expenditures required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5366}{website})
\item 3.2.7.10 - Advertising restricted on radio (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5367}{website})
\item 3.2.7.10 - Advertising restricted on television (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5368}{website})
\item 3.2.7.10 - Advertising restricted in print media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5369}{website})
\item 3.2.7.10 - Advertising restricted on the domestic internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5370}{website})
\item 3.2.7.10 - Advertising restricted on the global internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5371}{website})
\item 3.2.7.11 - Sponsorship of international events and activities restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5376}{website})
\item 3.2.7.11 - Tobacco sponsorship of participants therein restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5377}{website})
\item 3.2.7.12 - Cooperation on the elimination of cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5638}{website})
\item 3.2.7.13 - Penalties imposed for cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5639}{website})
\item 3.2.7.13 - Penalties imposed for cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5639}{website})
\item 3.2.7.14 - Progress made in implementing Article 13 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5640}{website})
\item 3.2.7.15 - Use of guidelines on Article 13 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7014}{website})
\item 3.2.7.16 - Additional information concerning tobacco advertising, promotion and sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5641}{website})
\item 3.2.7.2 - Ban on display of tobacco products at points of sales (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5349}{website})
\item 3.2.7.2 - Ban on display of tobacco products at points of sales (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5349}{website})
\item 3.2.7.2 - Ban covering product placement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5353}{website})
\item 3.2.7.2 - Ban covering product placement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5353}{website})
\item 3.2.7.2 - Ban covering the depiction/use of tobacco in entertainment media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5354}{website})
\item 3.2.7.2 - Ban covering the depiction/use of tobacco in entertainment media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5354}{website})
\item 3.2.7.2 - Ban covering tobacco sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5356}{website})
\item 3.2.7.2 - Ban covering tobacco sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5356}{website})
\item 3.2.7.13 - Penalties imposed for cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5639}{website})
\item 3.2.7.13 - Penalties imposed for cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5639}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art14}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
