% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art14}
\alias{art14}
\title{Article 14}
\format{A dataset with 26871 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art14
}
\description{
Article 14
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 3.2.8.1 - Evidence-based comprehensive and integrated guidelines developed (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5726}{website})
\item 3.2.8.2.a - Implemented media campaigns to promote tobacco cessation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5727}{website})
\item 3.2.8.2.c - Implemented programmes specially designed for women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5867}{website})
\item 3.2.8.2.d - Implemented programmes specially designed for pregnant women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5868}{website})
\item 3.2.8.2.b - Implemented programmes specially designed for underage girls and young women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5866}{website})
\item 3.2.8.2.e - Implemented telephone quitlines (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5869}{website})
\item 3.2.8.2.f - Implemented local events to promote cessation of tobacco use (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5729}{website})
\item 3.2.8.3.a - Designed programmes to promote cessation in educational institutions (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5100}{website})
\item 3.2.8.3.b - Designed programmes to promote cessation in health-care facilities (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5101}{website})
\item 3.2.8.3.c - Designed programmes to promote cessation in workplaces (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5102}{website})
\item 3.2.8.3.d - Designed programmes to promote cessation in sporting environments (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5103}{website})
\item 3.2.8.4.a - Included diagnosis and treatment in national tobacco control programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5731}{website})
\item 3.2.8.4.b - Included diagnosis and treatment in national health programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5732}{website})
\item 3.2.8.4.c - Included diagnosis and treatment in national educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5733}{website})
\item 3.2.8.5 - Included diagnosis and treatment in the health-care system (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5412}{website})
\item 3.2.8.6.a - Primary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5415}{website})
\item 3.2.8.6.a - Primary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5415}{website})
\item 3.2.8.6.b - Secondary and tertiary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5416}{website})
\item 3.2.8.6.b - Secondary and tertiary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5416}{website})
\item 3.2.8.6.c - Specialist health-care systems providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5417}{website})
\item 3.2.8.6.c - Specialist health-care systems providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5417}{website})
\item 3.2.8.6.d - Specialized centres for cessation providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5418}{website})
\item 3.2.8.6.d - Specialized centres for cessation providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5418}{website})
\item 3.2.8.6.e - Rehabilitation centres providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5419}{website})
\item 3.2.8.6.e - Rehabilitation centres providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5419}{website})
\item 3.2.8.7.a - Programmes in primary health care covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5646}{website})
\item 3.2.8.7.b - Programmes in secondary and tertiary health care covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5647}{website})
\item 3.2.8.7.c - Programmes in specialist health-care systems covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5648}{website})
\item 3.2.8.7.d - Programmes in specialized centers covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5651}{website})
\item 3.2.8.7.e - Programmes in rehabilitation centers covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5652}{website})
\item 3.2.8.8.a - Physicians offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5657}{website})
\item 3.2.8.8.b - Dentists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5658}{website})
\item 3.2.8.8.b - Dentists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5658}{website})
\item 3.2.8.8.c - Family doctors offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5659}{website})
\item 3.2.8.8.d - Traditional practitioners offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5660}{website})
\item 3.2.8.8.f - Nurses offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5662}{website})
\item 3.2.8.8.g - Midwives offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5663}{website})
\item 3.2.8.8.h - Pharmacists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5664}{website})
\item 3.2.8.8.i - Community workers offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5665}{website})
\item 3.2.8.8.j - Social workers offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5666}{website})
\item 3.2.8.9.a - Tobacco dependence treatment incorporated into the curricula of medical schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5812}{website})
\item 3.2.8.9.b - Tobacco dependence treatment incorporated into the curricula of dentist schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5813}{website})
\item 3.2.8.9.c - Tobacco dependence treatment incorporated into the curricula of nursing schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5814}{website})
\item 3.2.8.9.d - Tobacco dependence treatment incorporated into the curricula of pharmacy schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5817}{website})
\item 3.2.8.10 - Accessibility and affordability of pharmaceutical products facilitated (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5677}{website})
\item 3.2.8.11 - Where and how pharmaceutical products legally purchased (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5870}{website})
\item 3.2.8.12.a - Nicotine replacement therapy available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5679}{website})
\item 3.2.8.12.b - Treatment with bupropion available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5680}{website})
\item 3.2.8.12.b - Treatment with bupropion available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5680}{website})
\item 3.2.8.12.c - Treatment with varenicline available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5681}{website})
\item 3.2.8.14 - Progress made in implementing Article 14 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5742}{website})
\item 3.2.8.15 - Use of guidelines oin Article 14 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7015}{website})
\item 3.2.8.16 - Relevant information (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5743}{website})
\item 3.2.8.6.a - Primary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5415}{website})
\item 3.2.8.6.a - Primary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5415}{website})
\item 3.2.8.6.b - Secondary and tertiary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5416}{website})
\item 3.2.8.6.b - Secondary and tertiary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5416}{website})
\item 3.2.8.6.c - Specialist health-care systems providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5417}{website})
\item 3.2.8.6.c - Specialist health-care systems providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5417}{website})
\item 3.2.8.6.d - Specialized centres for cessation providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5418}{website})
\item 3.2.8.6.d - Specialized centres for cessation providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5418}{website})
\item 3.2.8.6.e - Rehabilitation centres providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5419}{website})
\item 3.2.8.6.e - Rehabilitation centres providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5419}{website})
\item 3.2.8.8.b - Dentists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5658}{website})
\item 3.2.8.8.b - Dentists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5658}{website})
\item 3.2.8.12.b - Treatment with bupropion available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5680}{website})
\item 3.2.8.12.b - Treatment with bupropion available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5680}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art15}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
