% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art15}
\alias{art15}
\title{Article 15}
\format{A dataset with 8619 rows and 4 columns.}
\usage{
art15
}
\description{
Article 15
}
\details{
This article includes the following sections:
\itemize{
\item 2.6.2 - Data on the percentage of smuggled tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5537}{website})
\item 2.6.3 - Percentage of national tobacco market consituted by smuggled products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5538}{website})
\item 3.3.1.1 - Marking that assists in determining the origin of product required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4417}{website})
\item 3.3.1.2 - Marking that assists in identifying legally sold products required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4418}{website})
\item 3.3.1.3 - Statement on all packages of tobacco products required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5694}{website})
\item 3.3.1.4 - Tracking regime to further secure the distribution system developed (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5695}{website})
\item 3.3.1.5 - Legible marking required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4419}{website})
\item 3.3.1.6 - Monitoring of cross-border trade required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5696}{website})
\item 3.3.1.7 - Information exchange facilitated (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5697}{website})
\item 3.3.1.8 - Legislation against illicit trade enacted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4420}{website})
\item 3.3.1.9 - Confiscated manufacturing equipment to be destroyed (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5698}{website})
\item 3.3.1.10 - Storage and distribution of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5699}{website})
\item 3.3.1.11 - Confiscation of proceeds derived from illicit trade enabled (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4421}{website})
\item 3.3.1.12 - Cooperation to eliminate illicit trade promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5702}{website})
\item 3.3.1.13 - Licensing required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4422}{website})
\item 3.3.1.14 - Progress made in implementing Article 15 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5700}{website})
\item 3.3.1.15 - Additional information concerning illicit trade in tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5701}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art16}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
