% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art16}
\alias{art16}
\title{Article 16}
\format{A dataset with 7605 rows and 4 columns.}
\usage{
art16
}
\description{
Article 16
}
\details{
This article includes the following sections:
\itemize{
\item 3.3.2.1 - Sales of tobacco products to minors prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4423}{website})
\item 3.3.2.1 - Minimum age for purchasing tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4496}{website})
\item 3.3.2.2 - Clear and prominent indicator required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5706}{website})
\item 3.3.2.3 - Required that sellers request for evidence of having reached full legal age (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5707}{website})
\item 3.3.2.4 - Ban of sale of tobacco in any directly accessible manner (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5708}{website})
\item 3.3.2.5 - Manufacture and sale of any objects in the form of tobacco products prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5709}{website})
\item 3.3.2.6 - Sale of tobacco products from vending machines prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5710}{website})
\item 3.3.2.7 - Tobacco vending machines not accessible to minors (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5711}{website})
\item 3.3.2.8.a - Distribution of free tobacco products to the public prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5712}{website})
\item 3.3.2.8.b - Distribution of free tobacco products to minors prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5713}{website})
\item 3.3.2.9 - Sale of cigarettes individually or in small packets prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4425}{website})
\item 3.3.2.10 - Penalties against sellers provided (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4426}{website})
\item 3.3.2.11 - Sale of tobacco products by minors prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4427}{website})
\item 3.3.2.12 - Progress made in implementing Article 16 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5714}{website})
\item 3.3.2.13 - Additional information concerning sales to and by minors (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5715}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art17}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
