% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art17}
\alias{art17}
\title{Article 17}
\format{A dataset with 4056 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art17
}
\description{
Article 17
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 2.7.1 - Tobacco growing in your jurisdiction (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5541}{website})
\item 2.7.2 - Number of workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5542}{website})
\item 2.7.3 - Share of the value of tobacco leaf (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5544}{website})
\item 3.3.3.1 - Viable alternatives for tobacco growers promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5081}{website})
\item 3.3.3.1 - Viable alternatives for tobacco workers promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5080}{website})
\item 3.3.3.1 - Viable alternatives for tobacco sellers promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5082}{website})
\item 3.3.3.2 - Progress made in implementing Article 17 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5724}{website})
\item 3.3.3.3 - All other information concerning provision of economically viable alternatives (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5725}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art16}}, \code{\link{art18}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
