% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art18}
\alias{art18}
\title{Article 18}
\format{A dataset with 3549 rows and 4 columns.}
\usage{
art18
}
\description{
Article 18
}
\details{
This article includes the following sections:
\itemize{
\item 3.4.1.1 - Measures implemented in respect to tobacco cultivation considering the protection of the environment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5467}{website})
\item 3.4.1.1 - Measures implemented in respect to tobacco cultivation considering the health of persons (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5468}{website})
\item 3.4.1.2 - Measures implemented in respect to tobacco manufacturing for the protection of the environment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5470}{website})
\item 3.4.1.2 - Measures implemented in respect to tobacco manufacturing considering the health of persons (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5471}{website})
\item 3.4.1.3 - Progress made in implementing Article 18 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5472}{website})
\item 3.4.1.3 - Progress made in implementing Article 18 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5472}{website})
\item 3.4.1.4 - Additional information concerning protection of the environment and the health of persons (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5473}{website})
\item 3.4.1.3 - Progress made in implementing Article 18 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5472}{website})
\item 3.4.1.3 - Progress made in implementing Article 18 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5472}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art16}}, \code{\link{art17}},
  \code{\link{art19}}, \code{\link{art20}}
}
\concept{FCTC articles}
\keyword{datasets}
