% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\docType{data}
\name{art20}
\alias{art20}
\title{Article 20}
\format{A dataset with 12168 rows and 4 columns.
\describe{
\item{party}{Character scalar. Name of the party member.}
\item{answer}{Character scalar. Response to the item.}
\item{year}{Integer scalar. Year of the report.}
\item{section}{Character scalar. Number and description of the section}
}}
\usage{
art20
}
\description{
Article 20
}
\details{
This article includes the following sections of the \link[=fctc-reporting]{Reporting Instrument}:
\itemize{
\item 3.4.3.1 a - Research on determinants of tobacco consumption promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5821}{website})
\item 3.4.3.1 b - Research on consequences of tobacco consumption promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5822}{website})
\item 3.4.3.1 c - Research on social and economic indicators promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5823}{website})
\item 3.4.3.1 d - Research on tobacco use among women promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5824}{website})
\item 3.4.3.1 e  - Research on exposure to tobacco smoke promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5085}{website})
\item 3.4.3.1 f - Research on identification of tobacco dependence treatment promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5825}{website})
\item 3.4.3.1 g - Research on alternative livelihoods promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5826}{website})
\item 3.4.3.2 - Training for those engaged in tobacco control provided (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5109}{website})
\item 3.4.3.3.a - National system for surveillance of patterns of tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5111}{website})
\item 3.4.3.3.b - National system for surveillance of determinants of tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5837}{website})
\item 3.4.3.3.c - National system for surveillance of consequences of tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5838}{website})
\item 3.4.3.3.e - National system for surveillance of exposure to tobacco smoke established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5839}{website})
\item 3.4.3.3.d - National system for surveillance of  indicators related to tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5112}{website})
\item 3.4.3.4 - List of surveys undertaken in the past (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5872}{website})
\item 3.4.3.5 - List of surveys that are planned to be repeated or undertaken (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5873}{website})
\item 3.4.3.5 - List of surveys that are planned to be repeated or undertaken (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5873}{website})
\item 3.4.3.6 - Scientific and technical information exchanged (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5114}{website})
\item 3.4.3.6 - Information on tobacco industry practices exchanged (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5115}{website})
\item 3.4.3.6 - Information on cultivation of tobacco exchanged (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5117}{website})
\item 3.4.3.7 - Database of laws and regulations on tobacco control established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5118}{website})
\item 3.4.3.7 - Database of information about the enforcement of laws established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5119}{website})
\item 3.4.3.7 - Database of the pertinent jurisprudence established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5120}{website})
\item 3.4.3.8 - Progress made in implementing Article 20 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5842}{website})
\item 3.4.3.9 - Additional information concerning research, surveillance and exchange of information (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5843}{website})
\item 3.4.3.5 - List of surveys that are planned to be repeated or undertaken (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5873}{website})
\item 3.4.3.5 - List of surveys that are planned to be repeated or undertaken (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5873}{website})
}
}
\seealso{
Other FCTC articles: \code{\link{art02_and_26_assistance_provided}},
  \code{\link{art02_and_26_assistance_received}},
  \code{\link{art02_and_26_other_questions_related_to_assistance}},
  \code{\link{art02_and_26_priorities_and_comments}},
  \code{\link{art05}}, \code{\link{art06}},
  \code{\link{art08}}, \code{\link{art09}},
  \code{\link{art10}}, \code{\link{art11}},
  \code{\link{art12}}, \code{\link{art13}},
  \code{\link{art14}}, \code{\link{art15}},
  \code{\link{art16}}, \code{\link{art17}},
  \code{\link{art18}}, \code{\link{art19}}
}
\concept{FCTC articles}
\keyword{datasets}
