% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctc.r
\name{fctc-treaty}
\alias{fctc-treaty}
\title{WHO Framework Convention on Tobacco Control}
\description{
WHO Framework Convention on Tobacco Control
}
\details{
The document, which is included as a vignette so you can open it using\preformatted{vignette("who-fctc-english", package = "fctc")
}
}
\references{
Latest version downloaded from
http://apps.who.int/iris/bitstream/10665/42811/1/9241591013.pdf?ua=1
}
