% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\name{Structure of the data}
\alias{Structure of the data}
\title{Data structure on the FCTC}
\description{
This entry describes how the data is structure in terms of articles of the
convention and sections of the reporting instrument.
}
\details{
The data is organized in the same way that it was presented on the website,
with the sections of the \link[=fctc-reporting]{Reporting Instrument} grouped by
articles.
}
\section{Article 2 And 26 Assistance Provided}{


This article is associated with the dataset \link{2_and_26_assistance_provided}. It includes the following sections:
\itemize{
\item 4.1 - On transfer of skills and technology (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4710}{website})
\item 4.2 - Expertise for tobacco control programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4713}{website})
\item 4.3 - Training and sensitisation of personnel (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4715}{website})
\item 4.4 - Equipment, supplies, logistics (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4717}{website})
\item 4.5 - Methods for tobacco control, e.g. treatment of nicotine addiction (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4719}{website})
\item 4.6 - Research on affordability (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5744}{website})
}
}

\section{Article 2 And 26 Assistance Received}{


This article is associated with the dataset \link{2_and_26_assistance_received}. It includes the following sections:
\itemize{
\item 4.1 - On transfer of skills and technology (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4712}{website})
\item 4.2 - Expertise for tobacco control programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4714}{website})
\item 4.3 - Training and sensitisation of personnel (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4716}{website})
\item 4.4 - Equipment, supplies, logistics (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4718}{website})
\item 4.5 - Methods for tobacco control, e.g. treatment of nicotine addiction (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4720}{website})
\item 4.6 - Research on affordability (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5834}{website})
}
}

\section{Article 2 And 26 Other Questions Related To Assistance}{


This article is associated with the dataset \link{2_and_26_other_questions_related_to_assistance}. It includes the following sections:
\itemize{
\item 4.7 - Identify the Party or Parties (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4729}{website})
\item 4.8 - Information about assistance (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5847}{website})
\item 4.9 - Assistance under consideration (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4727}{website})
\item 4.10 - Encouraging development institutions (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5844}{website})
\item 4.11 - Additional details on encouraging development institutions (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5845}{website})
}
}

\section{Article 2 And 26 Priorities And Comments}{


This article is associated with the dataset \link{2_and_26_priorities_and_comments}. It includes the following sections:
\itemize{
\item 5.1 - Priorities for implementation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4487}{website})
\item 5.2 - Specific gaps (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5745}{website})
\item 5.3 - Details on specific gaps (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5747}{website})
\item 5.4 - Constraints or barriers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4488}{website})
\item 5.5 - Any other relevant information (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4489}{website})
\item 5.6 - Suggestions on the reporting instrument (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5746}{website})
}
}

\section{Article 5}{


This article is associated with the dataset \link{5}. It includes the following sections:
\itemize{
\item 3.1.1.1 - Comprehensive multisectoral national tobacco control strategy (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5048}{website})
\item 3.1.1.2 - Tobacco control included in national health strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5049}{website})
\item 3.1.1.3 - Any aspect of tobacco referred in any national strategy (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5050}{website})
\item 3.1.1.4 - Focal point for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5784}{website})
\item 3.1.1.4 - Tobacco control unit (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5212}{website})
\item 3.1.1.4 - National coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5785}{website})
\item 3.1.1.5 - Details on focal point for tobacco control, tobacco control unit and national coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5214}{website})
\item 3.1.1.6 - Progress made in implementing Article 5.1 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5215}{website})
\item 3.1.1.7 - Additional information concerning general obligations (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5216}{website})
\item 3.1.2.1 - Interference by the tobacco industry (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5067}{website})
\item 3.1.2.2 - Public access to a wide range of information on the tobacco industry (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5790}{website})
\item 3.1.2.3 - Details on the protection of public health policies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5787}{website})
\item 3.1.2.4 - Progress made in protection of public health policies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5788}{website})
\item 3.1.2.5 - Use of guidelines on Article 5.3 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7004}{website})
\item 3.1.2.6 - Additional information concerning protection of public health policies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5789}{website})
\item 3.1.1.4 - Tobacco control unit (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5212}{website})
\item 3.1.1.4 - National coordinating mechanism for tobacco control (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5785}{website})
}
}

\section{Article 6}{


This article is associated with the dataset \link{6}. It includes the following sections:
\itemize{
\item 2.3.1 - Existence of information on tobacco-related mortality (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5528}{website})
\item 2.3.3 - Additional information on tobacco-related mortality (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5530}{website})
\item 2.4.1 - Existence of information on the economic burden of tobacco use (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5534}{website})
\item 2.8.1 - Proportion of the retail price consisting of taxes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5484}{website})
\item 2.8.2 - Only specific tax levied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5485}{website})
\item 2.8.2 - Only ad valorem tax levied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5486}{website})
\item 2.8.2 - Combination of specific and ad valorem taxes levied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5487}{website})
\item 2.8.2 - Complex tax structure applied (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5488}{website})
\item 2.8.4 - Trends in taxation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5502}{website})
\item 2.8.5 - Tobacco tax earmarking (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5503}{website})
\item 2.9.4 - Trends in prices (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5524}{website})
\item 3.2.1.1 - Tax policies to reduce tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5222}{website})
\item 3.2.1.2 - Tobacco sales to international travelers prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5223}{website})
\item 3.2.1.3 - Tobacco imports by international travelers prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5224}{website})
\item 3.2.1.4 - Progress made in implementing Article 6 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5225}{website})
\item 3.2.1.5 - Additional information concerning price and tax measures (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5226}{website})
\item 2.8.1 - Proportion of the retail price consisting of taxes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5484}{website})
}
}

\section{Article 8}{


This article is associated with the dataset \link{8}. It includes the following sections:
\itemize{
\item 2.2.1 - Availability of data on exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5525}{website})
\item 2.2.2 - Details concerning data on exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5526}{website})
\item 3.2.2.1 - Tobacco smoking banned in public places (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5026}{website})
\item 3.2.2.2 - National law providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5851}{website})
\item 3.2.2.2 - Subnational law(s) providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5852}{website})
\item 3.2.2.2 - Administrative and executive orders providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5853}{website})
\item 3.2.2.2 - Voluntary agreements providing for the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5854}{website})
\item 3.2.2.3 - Explanation of type/nature and content of the measures providing the ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5857}{website})
\item 3.2.2.4 - Mechanism/infrastructure for enforcement provided (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5858}{website})
\item 3.2.2.5 - Details of the mechanism/infrastructure for enforcement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5859}{website})
\item 3.2.2.8 - Progress made in implementing Article 8 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5570}{website})
\item 3.2.2.9 - Use of guidelines on Article 8 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7009}{website})
\item 3.2.2.10 - Additional information concerning protection from exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5571}{website})
\item 3.2.2.5 - Details of the mechanism/infrastructure for enforcement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5859}{website})
}
}

\section{Article 9}{


This article is associated with the dataset \link{9}. It includes the following sections:
\itemize{
\item 3.2.3.1 - Testing and measuring the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4430}{website})
\item 3.2.3.2 - Testing and measuring the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4431}{website})
\item 3.2.3.3 - Regulating the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4432}{website})
\item 3.2.3.4 - Regulating the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4433}{website})
\item 3.2.3.5 - Progress made in implementing Article 9 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5547}{website})
\item 3.2.3.6 - Use of guidelines on Article 9 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7011}{website})
\item 3.2.3.7 - Additional information concerning regulation of the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5548}{website})
}
}

\section{Article 10}{


This article is associated with the dataset \link{10}. It includes the following sections:
\itemize{
\item 3.2.4.1 - Requiring disclosure of information about the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5043}{website})
\item 3.2.4.1 - Requiring disclosure of information about the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5044}{website})
\item 3.2.4.2 - Requiring public disclosure on the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5791}{website})
\item 3.2.4.2 - Requiring public disclosure on the emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5792}{website})
\item 3.2.4.3 - Progress made in implementing Article 10 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5793}{website})
\item 3.2.4.5 - Additional information concerning regulation of tobacco product disclosures (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5794}{website})
\item 3.2.4.4 - Use of guidelines on Article 10 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7012}{website})
\item 3.2.4.2 - Requiring public disclosure on the contents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5791}{website})
}
}

\section{Article 11}{


This article is associated with the dataset \link{11}. It includes the following sections:
\itemize{
\item 3.2.5.1 - Packaging of tobacco products does not carry advertisement or promotion (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5575}{website})
\item 3.2.5.2 - Misleading descriptors (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4434}{website})
\item 3.2.5.3 - Health warnings required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4435}{website})
\item 3.2.5.4 - Health warnings approved by the competent national authority (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4436}{website})
\item 3.2.5.5 - Rotated health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4437}{website})
\item 3.2.5.6 - Large, clear, visible and legible health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4438}{website})
\item 3.2.5.8 - Health warnings occupying no less than 30% (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4439}{website})
\item 3.2.5.9 - Health warnings occupying 50% or more (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4440}{website})
\item 3.2.5.10 - Health warnings in the form of pictures or pictograms (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4441}{website})
\item 3.2.5.11 - Copyright to pictures owned by the Government (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5576}{website})
\item 3.2.5.12 - Granting of license for the use of health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5578}{website})
\item 3.2.5.13 - Constituents of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5874}{website})
\item 3.2.5.13 - Emissions of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5875}{website})
\item 3.2.5.15 - Progress made in implementing Article 11 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5579}{website})
\item 3.2.5.16 - Use of guidelines on Article 11 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7018}{website})
\item 3.2.5.17 - Additional information concerning packaging and labelling (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5580}{website})
\item 3.2.5.8 - Health warnings occupying no less than 30% (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4439}{website})
\item 3.2.5.9 - Health warnings occupying 50% or more (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4440}{website})
\item 3.2.5.12 - Granting of license for the use of health warnings (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5578}{website})
}
}

\section{Article 12}{


This article is associated with the dataset \link{12}. It includes the following sections:
\itemize{
\item 3.2.6.1 - Implemented educational and public awareness programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5069}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to adults (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5070}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to children and youth (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5071}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to men (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5583}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5585}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to pregnant women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5587}{website})
\item 3.2.6.2 - Implemented educational programmes targeted to ethnic groups (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5588}{website})
\item 3.2.6.3 - Age differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5594}{website})
\item 3.2.6.3 - Gender differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5595}{website})
\item 3.2.6.3 - Educational background differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5596}{website})
\item 3.2.6.3 - Cultural differences reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5597}{website})
\item 3.2.6.3  - Socioeconomic status reflected in educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5599}{website})
\item 3.2.6.4 - Programmes covering the health risks of tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5072}{website})
\item 3.2.6.4 - Programmes covering the risks of exposure to tobacco smoke (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5073}{website})
\item 3.2.6.4 - Programmes covering the benefits of cessation of tobacco use (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5074}{website})
\item 3.2.6.4 - Programmes covering economic consequences of tobacco production (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5601}{website})
\item 3.2.6.4 - Programmes covering economic consequences of tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5604}{website})
\item 3.2.6.4 - Programmes covering environmental consequences of tobacco production (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5606}{website})
\item 3.2.6.4 - Programmes covering environmental consequences of tobacco consumption (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5608}{website})
\item 3.2.6.5 - Public agencies involved in programmes and strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5611}{website})
\item 3.2.6.5 - NGOs involved in programmes and strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5612}{website})
\item 3.2.6.5 - Private organizations involved in programmes and strategies (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5615}{website})
\item 3.2.6.6 - Programmes guided by research (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5618}{website})
\item 3.2.6.7 - Training programmes addressed to health workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5090}{website})
\item 3.2.6.7 - Training programmes addressed to community workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5091}{website})
\item 3.2.6.7 - Training programmes addressed to social workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5092}{website})
\item 3.2.6.7 - Training programmes addressed to media professionals (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5093}{website})
\item 3.2.6.7 - Training programmes addressed to educators (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5094}{website})
\item 3.2.6.7 - Training programmes addressed to decision-makers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5095}{website})
\item 3.2.6.7 - Training programmes addressed to administrators (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5096}{website})
\item 3.2.6.8 - Progress made in implementing Article 12 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5621}{website})
\item 3.2.6.9 - Use of guidelines on Article 12 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7013}{website})
\item 3.2.6.10 - Additional information concerning education and communication (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5622}{website})
}
}

\section{Article 13}{


This article is associated with the dataset \link{13}. It includes the following sections:
\itemize{
\item 3.2.7.1 - Comprehensive ban on all tobacco advertising, promotion and sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5346}{website})
\item 3.2.7.2 - Ban on display of tobacco products at points of sales (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5349}{website})
\item 3.2.7.2 - Ban covering the domestic internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5350}{website})
\item 3.2.7.2 - Ban covering the global internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5351}{website})
\item 3.2.7.2 - Ban covering brand stretching and/or sharing (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5352}{website})
\item 3.2.7.2 - Ban covering product placement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5353}{website})
\item 3.2.7.2 - Ban covering the depiction/use of tobacco in entertainment media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5354}{website})
\item 3.2.7.2 - Ban covering tobacco sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5356}{website})
\item 3.2.7.2 - Ban covering corporate social responsibility (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5357}{website})
\item 3.2.7.2 - Ban covering cross-border advertising originating from the country (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5358}{website})
\item 3.2.7.2 - Ban covering cross-border advertising entering the country (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5359}{website})
\item 3.2.7.3 - Precluded by constitution from undertaking a comprehensive ban (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5360}{website})
\item 3.2.7.4 - All tobacco advertising, promotion and sponsorship restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5361}{website})
\item 3.2.7.5 - Cross-border advertising originating from the country restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5362}{website})
\item 3.2.7.6 - Advertising by false and misleading means prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5363}{website})
\item 3.2.7.7 - Use of warnings to accompany all advertising required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5364}{website})
\item 3.2.7.8 - Use of direct or indirect incentives restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5365}{website})
\item 3.2.7.9 - Disclosure of advertising expenditures required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5366}{website})
\item 3.2.7.10 - Advertising restricted on radio (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5367}{website})
\item 3.2.7.10 - Advertising restricted on television (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5368}{website})
\item 3.2.7.10 - Advertising restricted in print media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5369}{website})
\item 3.2.7.10 - Advertising restricted on the domestic internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5370}{website})
\item 3.2.7.10 - Advertising restricted on the global internet (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5371}{website})
\item 3.2.7.11 - Sponsorship of international events and activities restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5376}{website})
\item 3.2.7.11 - Tobacco sponsorship of participants therein restricted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5377}{website})
\item 3.2.7.12 - Cooperation on the elimination of cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5638}{website})
\item 3.2.7.13 - Penalties imposed for cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5639}{website})
\item 3.2.7.14 - Progress made in implementing Article 13 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5640}{website})
\item 3.2.7.15 - Use of guidelines on Article 13 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7014}{website})
\item 3.2.7.16 - Additional information concerning tobacco advertising, promotion and sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5641}{website})
\item 3.2.7.2 - Ban on display of tobacco products at points of sales (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5349}{website})
\item 3.2.7.2 - Ban covering product placement (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5353}{website})
\item 3.2.7.2 - Ban covering the depiction/use of tobacco in entertainment media (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5354}{website})
\item 3.2.7.2 - Ban covering tobacco sponsorship (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5356}{website})
\item 3.2.7.13 - Penalties imposed for cross-border advertising (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5639}{website})
}
}

\section{Article 14}{


This article is associated with the dataset \link{14}. It includes the following sections:
\itemize{
\item 3.2.8.1 - Evidence-based comprehensive and integrated guidelines developed (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5726}{website})
\item 3.2.8.2.a - Implemented media campaigns to promote tobacco cessation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5727}{website})
\item 3.2.8.2.c - Implemented programmes specially designed for women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5867}{website})
\item 3.2.8.2.d - Implemented programmes specially designed for pregnant women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5868}{website})
\item 3.2.8.2.b - Implemented programmes specially designed for underage girls and young women (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5866}{website})
\item 3.2.8.2.e - Implemented telephone quitlines (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5869}{website})
\item 3.2.8.2.f - Implemented local events to promote cessation of tobacco use (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5729}{website})
\item 3.2.8.3.a - Designed programmes to promote cessation in educational institutions (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5100}{website})
\item 3.2.8.3.b - Designed programmes to promote cessation in health-care facilities (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5101}{website})
\item 3.2.8.3.c - Designed programmes to promote cessation in workplaces (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5102}{website})
\item 3.2.8.3.d - Designed programmes to promote cessation in sporting environments (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5103}{website})
\item 3.2.8.4.a - Included diagnosis and treatment in national tobacco control programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5731}{website})
\item 3.2.8.4.b - Included diagnosis and treatment in national health programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5732}{website})
\item 3.2.8.4.c - Included diagnosis and treatment in national educational programmes (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5733}{website})
\item 3.2.8.5 - Included diagnosis and treatment in the health-care system (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5412}{website})
\item 3.2.8.6.a - Primary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5415}{website})
\item 3.2.8.6.b - Secondary and tertiary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5416}{website})
\item 3.2.8.6.c - Specialist health-care systems providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5417}{website})
\item 3.2.8.6.d - Specialized centres for cessation providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5418}{website})
\item 3.2.8.6.e - Rehabilitation centres providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5419}{website})
\item 3.2.8.7.a - Programmes in primary health care covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5646}{website})
\item 3.2.8.7.b - Programmes in secondary and tertiary health care covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5647}{website})
\item 3.2.8.7.c - Programmes in specialist health-care systems covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5648}{website})
\item 3.2.8.7.d - Programmes in specialized centers covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5651}{website})
\item 3.2.8.7.e - Programmes in rehabilitation centers covered by public funding (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5652}{website})
\item 3.2.8.8.a - Physicians offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5657}{website})
\item 3.2.8.8.b - Dentists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5658}{website})
\item 3.2.8.8.c - Family doctors offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5659}{website})
\item 3.2.8.8.d - Traditional practitioners offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5660}{website})
\item 3.2.8.8.f - Nurses offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5662}{website})
\item 3.2.8.8.g - Midwives offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5663}{website})
\item 3.2.8.8.h - Pharmacists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5664}{website})
\item 3.2.8.8.i - Community workers offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5665}{website})
\item 3.2.8.8.j - Social workers offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5666}{website})
\item 3.2.8.9.a - Tobacco dependence treatment incorporated into the curricula of medical schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5812}{website})
\item 3.2.8.9.b - Tobacco dependence treatment incorporated into the curricula of dentist schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5813}{website})
\item 3.2.8.9.c - Tobacco dependence treatment incorporated into the curricula of nursing schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5814}{website})
\item 3.2.8.9.d - Tobacco dependence treatment incorporated into the curricula of pharmacy schools (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5817}{website})
\item 3.2.8.10 - Accessibility and affordability of pharmaceutical products facilitated (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5677}{website})
\item 3.2.8.11 - Where and how pharmaceutical products legally purchased (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5870}{website})
\item 3.2.8.12.a - Nicotine replacement therapy available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5679}{website})
\item 3.2.8.12.b - Treatment with bupropion available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5680}{website})
\item 3.2.8.12.c - Treatment with varenicline available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5681}{website})
\item 3.2.8.14 - Progress made in implementing Article 14 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5742}{website})
\item 3.2.8.15 - Use of guidelines oin Article 14 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=7015}{website})
\item 3.2.8.16 - Relevant information (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5743}{website})
\item 3.2.8.6.a - Primary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5415}{website})
\item 3.2.8.6.b - Secondary and tertiary health care providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5416}{website})
\item 3.2.8.6.c - Specialist health-care systems providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5417}{website})
\item 3.2.8.6.d - Specialized centres for cessation providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5418}{website})
\item 3.2.8.6.e - Rehabilitation centres providing programmes on diagnosis and treatment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5419}{website})
\item 3.2.8.8.b - Dentists offering counseling services (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5658}{website})
\item 3.2.8.12.b - Treatment with bupropion available (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5680}{website})
}
}

\section{Article 15}{


This article is associated with the dataset \link{15}. It includes the following sections:
\itemize{
\item 2.6.2 - Data on the percentage of smuggled tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5537}{website})
\item 2.6.3 - Percentage of national tobacco market consituted by smuggled products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5538}{website})
\item 3.3.1.1 - Marking that assists in determining the origin of product required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4417}{website})
\item 3.3.1.2 - Marking that assists in identifying legally sold products required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4418}{website})
\item 3.3.1.3 - Statement on all packages of tobacco products required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5694}{website})
\item 3.3.1.4 - Tracking regime to further secure the distribution system developed (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5695}{website})
\item 3.3.1.5 - Legible marking required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4419}{website})
\item 3.3.1.6 - Monitoring of cross-border trade required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5696}{website})
\item 3.3.1.7 - Information exchange facilitated (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5697}{website})
\item 3.3.1.8 - Legislation against illicit trade enacted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4420}{website})
\item 3.3.1.9 - Confiscated manufacturing equipment to be destroyed (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5698}{website})
\item 3.3.1.10 - Storage and distribution of tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5699}{website})
\item 3.3.1.11 - Confiscation of proceeds derived from illicit trade enabled (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4421}{website})
\item 3.3.1.12 - Cooperation to eliminate illicit trade promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5702}{website})
\item 3.3.1.13 - Licensing required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4422}{website})
\item 3.3.1.14 - Progress made in implementing Article 15 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5700}{website})
\item 3.3.1.15 - Additional information concerning illicit trade in tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5701}{website})
}
}

\section{Article 16}{


This article is associated with the dataset \link{16}. It includes the following sections:
\itemize{
\item 3.3.2.1 - Sales of tobacco products to minors prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4423}{website})
\item 3.3.2.1 - Minimum age for purchasing tobacco products (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4496}{website})
\item 3.3.2.2 - Clear and prominent indicator required (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5706}{website})
\item 3.3.2.3 - Required that sellers request for evidence of having reached full legal age (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5707}{website})
\item 3.3.2.4 - Ban of sale of tobacco in any directly accessible manner (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5708}{website})
\item 3.3.2.5 - Manufacture and sale of any objects in the form of tobacco products prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5709}{website})
\item 3.3.2.6 - Sale of tobacco products from vending machines prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5710}{website})
\item 3.3.2.7 - Tobacco vending machines not accessible to minors (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5711}{website})
\item 3.3.2.8.a - Distribution of free tobacco products to the public prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5712}{website})
\item 3.3.2.8.b - Distribution of free tobacco products to minors prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5713}{website})
\item 3.3.2.9 - Sale of cigarettes individually or in small packets prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4425}{website})
\item 3.3.2.10 - Penalties against sellers provided (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4426}{website})
\item 3.3.2.11 - Sale of tobacco products by minors prohibited (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=4427}{website})
\item 3.3.2.12 - Progress made in implementing Article 16 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5714}{website})
\item 3.3.2.13 - Additional information concerning sales to and by minors (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5715}{website})
}
}

\section{Article 17}{


This article is associated with the dataset \link{17}. It includes the following sections:
\itemize{
\item 2.7.1 - Tobacco growing in your jurisdiction (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5541}{website})
\item 2.7.2 - Number of workers (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5542}{website})
\item 2.7.3 - Share of the value of tobacco leaf (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5544}{website})
\item 3.3.3.1 - Viable alternatives for tobacco growers promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5081}{website})
\item 3.3.3.1 - Viable alternatives for tobacco workers promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5080}{website})
\item 3.3.3.1 - Viable alternatives for tobacco sellers promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5082}{website})
\item 3.3.3.2 - Progress made in implementing Article 17 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5724}{website})
\item 3.3.3.3 - All other information concerning provision of economically viable alternatives (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5725}{website})
}
}

\section{Article 18}{


This article is associated with the dataset \link{18}. It includes the following sections:
\itemize{
\item 3.4.1.1 - Measures implemented in respect to tobacco cultivation considering the protection of the environment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5467}{website})
\item 3.4.1.1 - Measures implemented in respect to tobacco cultivation considering the health of persons (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5468}{website})
\item 3.4.1.2 - Measures implemented in respect to tobacco manufacturing for the protection of the environment (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5470}{website})
\item 3.4.1.2 - Measures implemented in respect to tobacco manufacturing considering the health of persons (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5471}{website})
\item 3.4.1.3 - Progress made in implementing Article 18 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5472}{website})
\item 3.4.1.4 - Additional information concerning protection of the environment and the health of persons (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5473}{website})
\item 3.4.1.3 - Progress made in implementing Article 18 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5472}{website})
}
}

\section{Article 19}{


This article is associated with the dataset \link{19}. It includes the following sections:
\itemize{
\item 3.4.2.1 - Measures on criminal liability contained in the tobacco control legislation (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5876}{website})
\item 3.4.2.2 - Separate liability provisions on tobacco control outside of the tobacco control legislation exist (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5877}{website})
\item 3.4.2.3 - Civil liability measures that are specific to tobacco control exist (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5878}{website})
\item 3.4.2.4 - Civil liability measures that could apply to tobacco control exist (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5879}{website})
}
}

\section{Article 20}{


This article is associated with the dataset \link{20}. It includes the following sections:
\itemize{
\item 3.4.3.1 a - Research on determinants of tobacco consumption promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5821}{website})
\item 3.4.3.1 b - Research on consequences of tobacco consumption promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5822}{website})
\item 3.4.3.1 c - Research on social and economic indicators promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5823}{website})
\item 3.4.3.1 d - Research on tobacco use among women promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5824}{website})
\item 3.4.3.1 e  - Research on exposure to tobacco smoke promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5085}{website})
\item 3.4.3.1 f - Research on identification of tobacco dependence treatment promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5825}{website})
\item 3.4.3.1 g - Research on alternative livelihoods promoted (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5826}{website})
\item 3.4.3.2 - Training for those engaged in tobacco control provided (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5109}{website})
\item 3.4.3.3.a - National system for surveillance of patterns of tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5111}{website})
\item 3.4.3.3.b - National system for surveillance of determinants of tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5837}{website})
\item 3.4.3.3.c - National system for surveillance of consequences of tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5838}{website})
\item 3.4.3.3.e - National system for surveillance of exposure to tobacco smoke established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5839}{website})
\item 3.4.3.3.d - National system for surveillance of  indicators related to tobacco consumption established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5112}{website})
\item 3.4.3.4 - List of surveys undertaken in the past (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5872}{website})
\item 3.4.3.5 - List of surveys that are planned to be repeated or undertaken (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5873}{website})
\item 3.4.3.6 - Scientific and technical information exchanged (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5114}{website})
\item 3.4.3.6 - Information on tobacco industry practices exchanged (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5115}{website})
\item 3.4.3.6 - Information on cultivation of tobacco exchanged (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5117}{website})
\item 3.4.3.7 - Database of laws and regulations on tobacco control established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5118}{website})
\item 3.4.3.7 - Database of information about the enforcement of laws established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5119}{website})
\item 3.4.3.7 - Database of the pertinent jurisprudence established (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5120}{website})
\item 3.4.3.8 - Progress made in implementing Article 20 (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5842}{website})
\item 3.4.3.9 - Additional information concerning research, surveillance and exchange of information (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5843}{website})
\item 3.4.3.5 - List of surveys that are planned to be repeated or undertaken (\href{http://apps.who.int/fctc/implementation/database/sites/implementation/scripts/src/tabulardata.php?indicator=5873}{website})
}
}

\section{Last update}{

This dataset has been generated on 2017-12-01 11:52:53
}

